% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_MLR.R
\name{analyze_residuals}
\alias{analyze_residuals}
\title{Analyze Residuals from Multiple Linear Regression}
\usage{
analyze_residuals(df, dep_var, ind_vars)
}
\arguments{
\item{df}{A data frame.}

\item{dep_var}{A string of the dependent variable name.}

\item{ind_vars}{A vector of strings of the independent variable names.}
}
\value{
A list containing:
  \item{reduced}{The reduced model summary}
  \item{full}{The full model summary}
  \item{reduced_AIC}{AIC for the reduced model}
  \item{full_AIC}{AIC for the full model}
  \item{reduced_plots}{A list of plots for the reduced model}
  \item{full_plots}{A list of plots for the full model}
}
\description{
This function takes a data-set and constructs both a reduced linear model
(without interactions) and a full model (with all two-way interactions).
Residual plots and model summaries for both models are generated.
}
\examples{
data <- data.frame(Y = rnorm(50, mean = 50, sd = 10),
X1 = rnorm(50, mean = 5, sd = 2),
X2 = sample(c("A", "B", "C"), 50, replace = TRUE))
analyze_residuals(data, "Y", c("X1", "X2"))
}
